Move required and necessityIndicators from Label to FieldLabel
