+++ packages/components/src/contextual-help/src/ContextualHelp.module.css
@@ -15,11 +15,13 @@
     --hop-ContextualHelp-column-gap: 0;
     --hop-ContextualHelp-align-items: center;
     --hop-ContextualHelp-justify-content: center;
+    --hop-ContextualHelp-order: calc(infinity);
 
     display: var(--hop-ContextualHelp-display);
     column-gap: var(--hop-ContextualHelp-column-gap);
     align-items: var(--hop-ContextualHelp-align-items);
     justify-content: var(--hop-ContextualHelp-justify-content);
+    order: calc(infinity);
 
     box-sizing: var(--hop-ContextualHelp-box-sizing);
     aspect-ratio: var(--hop-ContextualHelp-aspect-ratio);
@@ -53,7 +55,6 @@
     --hop-ContextualHelp__popover-font-smoothing: antialiased;
     --hop-ContextualHelp__popover-color: var(--hop-neutral-text-strong);
     --hop-ContextualHelp__popover-padding: var(--hop-space-inset-squish-md);
-    --hop-ContextualHelp__popover-translate-amount: 0.5rem;
 
     max-inline-size: min(var(--hop-ContextualHelp__popover-max-inline-size), calc(100% - (var(--container-padding) * 2)));
     padding: var(--hop-ContextualHelp__popover-padding);
@@ -70,32 +71,8 @@
     border-radius: var(--hop-ContextualHelp__popover-border-radius);
     outline: var(--hop-ContextualHelp__popover-outline);
     box-shadow: var(--hop-ContextualHelp__popover-box-shadow);
-
-    transition: transform 0.2s, opacity 0.2s;
 }
 
 .hop-ContextualHelp__popover:focus-visible {
 	outline: var(--hop-ContextualHelp__popover-outline);
-}
-
-.hop-ContextualHelp__popover[data-entering],
-.hop-ContextualHelp__popover[data-exiting] {
-    transform: var(--origin);
-    opacity: 0;
-}
-
-.hop-ContextualHelp__popover[data-placement="top"] {
-    --origin: translateY(var(--hop-ContextualHelp__popover-translate-amount));
-}
-
-.hop-ContextualHelp__popover[data-placement="bottom"] {
-    --origin: translateY(calc(-1 * var(--hop-ContextualHelp__popover-translate-amount)));
-}
-
-.hop-ContextualHelp__popover[data-placement="right"] {
-    --origin: translateX(calc(-1 * var(--hop-ContextualHelp__popover-translate-amount)));
-}
-
-.hop-ContextualHelp__popover[data-placement="left"] {
-    --origin: translateX(var(--hop-ContextualHelp__popover-translate-amount));
 }
