+++ packages/components/src/typography/label/docs/optional.tsx
@@ -1,7 +1,7 @@
-import { Label } from "@hopper-ui/components";
+import { FieldLabel } from "@hopper-ui/components";
 
 export default function Example() {
     return (
-        <Label necessityIndicator="label">Username</Label>
+        <FieldLabel necessityIndicator="label">Username</FieldLabel>
     );
 }
