+++ packages/components/src/typography/label/tests/chromatic/FieldLabel.stories.tsx
@@ -37,3 +37,4 @@ export const Optional: Story = {
         necessityIndicator: "label"
     }
 };
+
